chore: refactor bulk ingest utils
